Resolved lingering win/posix cross-platform issue
The prior commit included a typo that prevented mapping of stdout.  This is resolved in this commit.

In addition, there is a lingering issue detected in testing in which the test windows system's calls to Popen were not converting the input list to a string for execution.  This commit explicitly checks for windows systems, in __build_command__ and outputs a string on windows systems, and a list on non-windows systems.