Use exceptions and return values, not state and error codes.
Note: Significant change to the interface. Updated version number in line with semantic versioning.

The error code is no longer stored in the object. Either a return value is immediately given or an exception is immediately raised.

Most adb calls are now assumed to not have output to stderr and will raise exceptions if they do. If this assumption is not accurate, they will need to change to return both output and error, like shell_command does.

I no longer trust whatsapp.py to behave correctly along different paths. Attempted to make it the same, but can't test it.